Refactor: 컴포넌트 기능에 맞게 구조 변경
