+++ src/_zlib/components/tweet/Tweet.tsx
@@ -1,11 +1,11 @@
-import { Icon } from './common/Icon';
-import { Menu } from './common/Menu';
-import { Spacing } from './common/Spacing';
-import { Txt } from './common/Txt';
-import { Wrapper } from './common/Wrapper';
-import colors from '../constants/colors';
-import { TweetType } from '../hooks/useTweets';
-import { db, storage } from '../server/firebase';
+import { Icon } from '../common/Icon';
+import { Menu } from '../common/Menu';
+import { Spacing } from '../common/Spacing';
+import { Txt } from '../common/Txt';
+import { Wrapper } from '../common/Wrapper';
+import colors from '../../constants/colors';
+import { TweetType } from '../../hooks/useTweets';
+import { db, storage } from '../../server/firebase';
 import { User } from 'firebase/auth';
 import { deleteDoc, doc, updateDoc } from 'firebase/firestore';
 import { deleteObject, ref } from 'firebase/storage';
