+++ src/pages/Home.tsx
@@ -1,7 +1,7 @@
 import { Wrapper } from '../_zlib/components/common/Wrapper';
-import PostTweet from '../_zlib/components/PostTweet';
-import TimeLine from '../_zlib/components/TimeLine';
-import Title from '../_zlib/components/Title';
+import PostTweet from '../_zlib/components/tweet/PostTweet';
+import TimeLine from '../_zlib/components/tweet/TimeLine';
+import Title from '../_zlib/components/common/Title';
 import useTweets from '../_zlib/hooks/useTweets';
 
 export default function Home() {
