+++ src/pages/Profile.tsx
@@ -4,8 +4,8 @@ import { Input } from '../_zlib/components/common/Input';
 import { Spacing } from '../_zlib/components/common/Spacing';
 import { Txt } from '../_zlib/components/common/Txt';
 import { Wrapper } from '../_zlib/components/common/Wrapper';
-import TimeLine from '../_zlib/components/TimeLine';
-import Title from '../_zlib/components/Title';
+import TimeLine from '../_zlib/components/tweet/TimeLine';
+import Title from '../_zlib/components/common/Title';
 import colors from '../_zlib/constants/colors';
 import useProfile from '../_zlib/hooks/useProfile';
 import useTweets from '../_zlib/hooks/useTweets';
